requirements.txt
torch
torchvision
numpy
opencv-python
matplotlib
pandas
scikit-learn
scipy
Pillow